Add LaTeX version of SRD
Added a LaTeX version of the document. This was _mostly_ generated from the markdown using some regex, with a second pass manually to correct errors and tweak bits.

I'm currently in the process of writing my own forged in the dark game, which I plan on laying out in LaTeX, so I thought since I'll need this content, perhaps other people could benefit from it too!